+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [DEV]
+- removed deprecated code: AbstractTemplate, AbstractFramework, ...
+- twig template base logic moved into contao-twig-support-bundle
+- replaced parseTemplate and parseWidget with BeforeParseTwigTemplateEvent and BeforeRenderTwigTemplateEvent
+- deprecated BeforeRenderTwigTemplateEvent
+- added layout to BeforeRenderCallback and PrepareTemplateCallback
+- removed legacyTemplate from BeforeRenderCallback and PrepareTemplateCallback
+- code enhancements
+
 ## [2.0.0-beta7-DEV] - 2020-09-15
 - added errorClass variable to multiple module templates
 - added error block to multiple variable templates